pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcripty"
version = "0.2.0"
description = "Standalone audio transcription and speaker diarization package"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Samuel Willems" },
]
dependencies = [
    "faster-whisper>=1.0.0",
    "pydub>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
    "audioop-lts>=0.2.1; python_version>='3.13'",
]

[project.optional-dependencies]
diarization = [
    "pyannote.audio>=3.1.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
transcripty = "transcripty.cli:cli"

[project.urls]
Homepage = "https://github.com/smwlms/transcripty"
Repository = "https://github.com/smwlms/transcripty"
Issues = "https://github.com/smwlms/transcripty/issues"

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
